Create ROADMAP
Project Mission & Summary
We are building The Open Body Database, an open-source platform for sharing human body data so that we can support a transitioned sustainable fashion industry, research and design based on human body realities. Dealing with the data of the human form requires openness in communication and as every human form is unique, the benefits of utilizing those unique variables within a database and within the community are substantial. Personal preference of privacy, gender discussions, forms, shapes and the multitude of variables the human form contains must be built from a trusting community with body positivity outlook and acceptance of all preferences. Furthermore the data collection and privacy forms must reflect and adhere to these policies.